finish splitting entity inference into two steps
